… Orm/Xtensive.Orm.Tests.Core/Modelling/IndexingModel/PrimaryIndexInfo.cs
… Orm/Xtensive.Orm/Orm/Internals/KeyFactory.cs
… Orm/Xtensive.Orm/Orm/Linq/ItemToTupleConverter{TItem}.cs
… Orm/Xtensive.Orm/Orm/Linq/QueryHelper.cs
… Orm/Xtensive.Orm/Orm/Model/IndexInfo.cs
… Orm/Xtensive.Orm/Orm/Model/TypeInfo.cs
… Orm/Xtensive.Orm/Orm/Rse/ColumnCollection.cs
… Orm/Xtensive.Orm/Orm/Rse/ColumnGroupCollection.cs
… Orm/Xtensive.Orm/Orm/Rse/Providers/Compilable/AggregateProvider.cs
… Orm/Xtensive.Orm/Orm/Rse/Providers/Compilable/ContainsTableProvider.cs
… Orm/Xtensive.Orm/Orm/Rse/Providers/Compilable/FreeTextProvider.cs
… Orm/Xtensive.Orm/Orm/Rse/Providers/Compilable/IncludeProvider.cs
… Orm/Xtensive.Orm/Orm/Rse/Providers/Compilable/OrderProviderBase.cs
… Orm/Xtensive.Orm/Orm/Rse/RecordSetHeader.cs
… Orm/Xtensive.Orm/Tuples/AccessorDelegates.cs
… Orm/Xtensive.Orm/Tuples/Packed/PackedFieldAccessor.cs
… Orm/Xtensive.Orm/Tuples/Packed/PackedFieldAccessorFactory.cs
… Orm/Xtensive.Orm/Tuples/Packed/PackedFieldDescriptor.cs
… Orm/Xtensive.Orm/Tuples/Packed/PackedFieldDescriptorComparer.cs
… Orm/Xtensive.Orm/Tuples/Packed/PackedTuple.cs
… Orm/Xtensive.Orm/Tuples/Packed/PackingOrderInfoComparer.cs
… Orm/Xtensive.Orm/Tuples/Packed/TupleLayout.cs
… Orm/Xtensive.Orm/Tuples/Tuple.cs
… Orm/Xtensive.Orm/Tuples/TupleDescriptor.cs
… Orm/Xtensive.Orm/Tuples/TupleExtensions.cs
… Orm/Xtensive.Orm/Xtensive.Orm.csproj
… Version.props